requirements.txt
flask
google-cloud-vision
python-docx
sentence-transformers
torch
